+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Autosave form not discarding autosave
 - Summary and Calculation Elements not showing on read only forms
+- number input not allowing zeros
 
 ## [2.5.2] - 2023-01-16